Move UI components to components directory
