--- frontend/src/components/ui/Loading.vue
@@ -1,16 +1,0 @@
-<template>
-	<div class="loading">
-		<loading-icon />
-	</div>
-</template>
-
-<script>
-import LoadingIcon from "../../assets/images/icons/loading";
-
-export default {
-	name: "Loading",
-	components: {
-		LoadingIcon
-	}
-};
-</script>
